\name{cf_wrappers}
\alias{cf_highcharts_scatter}
\alias{cf_highcharts_line}
\alias{cf_highcharts_spline}
\alias{cf_highcharts_step}
\alias{cf_highcharts_bar}
\alias{cf_highcharts_stacked_bar}
\alias{cf_highcharts_area}
\alias{cf_highcharts_stacked_area}
\alias{cf_highcharts_histogram}
\alias{cf_highcharts_density}
\alias{cf_highcharts_pie}
\alias{cf_highcharts_donut}
\alias{cf_highcharts_bubble}
\alias{cf_highcharts_boxplot}
\alias{cf_highcharts_heatmap}
\alias{cf_highcharts_treemap}
\alias{cf_highcharts_lollipop}
\alias{cf_plotly_scatter}
\alias{cf_plotly_line}
\alias{cf_plotly_spline}
\alias{cf_plotly_step}
\alias{cf_plotly_bar}
\alias{cf_plotly_stacked_bar}
\alias{cf_plotly_area}
\alias{cf_plotly_stacked_area}
\alias{cf_plotly_histogram}
\alias{cf_plotly_density}
\alias{cf_plotly_pie}
\alias{cf_plotly_donut}
\alias{cf_plotly_bubble}
\alias{cf_plotly_boxplot}
\alias{cf_plotly_heatmap}
\alias{cf_plotly_treemap}
\alias{cf_plotly_lollipop}
\alias{cf_d3_scatter}
\alias{cf_d3_line}
\alias{cf_d3_spline}
\alias{cf_d3_step}
\alias{cf_d3_bar}
\alias{cf_d3_stacked_bar}
\alias{cf_d3_area}
\alias{cf_d3_stacked_area}
\alias{cf_d3_histogram}
\alias{cf_d3_density}
\alias{cf_d3_pie}
\alias{cf_d3_donut}
\alias{cf_d3_bubble}
\alias{cf_d3_boxplot}
\alias{cf_d3_heatmap}
\alias{cf_d3_treemap}
\alias{cf_d3_lollipop}
\alias{cf_vega_scatter}
\alias{cf_vega_line}
\alias{cf_vega_spline}
\alias{cf_vega_step}
\alias{cf_vega_bar}
\alias{cf_vega_stacked_bar}
\alias{cf_vega_area}
\alias{cf_vega_stacked_area}
\alias{cf_vega_histogram}
\alias{cf_vega_density}
\alias{cf_vega_pie}
\alias{cf_vega_donut}
\alias{cf_vega_bubble}
\alias{cf_vega_boxplot}
\alias{cf_vega_heatmap}
\alias{cf_vega_treemap}
\alias{cf_vega_lollipop}
\alias{cf_observable_scatter}
\alias{cf_observable_line}
\alias{cf_observable_spline}
\alias{cf_observable_step}
\alias{cf_observable_bar}
\alias{cf_observable_stacked_bar}
\alias{cf_observable_area}
\alias{cf_observable_stacked_area}
\alias{cf_observable_histogram}
\alias{cf_observable_density}
\alias{cf_observable_pie}
\alias{cf_observable_donut}
\alias{cf_observable_bubble}
\alias{cf_observable_boxplot}
\alias{cf_observable_heatmap}
\alias{cf_observable_treemap}
\alias{cf_observable_lollipop}
\alias{cf_chartjs_scatter}
\alias{cf_chartjs_line}
\alias{cf_chartjs_spline}
\alias{cf_chartjs_step}
\alias{cf_chartjs_bar}
\alias{cf_chartjs_stacked_bar}
\alias{cf_chartjs_area}
\alias{cf_chartjs_stacked_area}
\alias{cf_chartjs_histogram}
\alias{cf_chartjs_density}
\alias{cf_chartjs_pie}
\alias{cf_chartjs_donut}
\alias{cf_chartjs_bubble}
\alias{cf_chartjs_boxplot}
\alias{cf_chartjs_heatmap}
\alias{cf_chartjs_treemap}
\alias{cf_chartjs_lollipop}
\alias{cf_echarts_scatter}
\alias{cf_echarts_line}
\alias{cf_echarts_spline}
\alias{cf_echarts_step}
\alias{cf_echarts_bar}
\alias{cf_echarts_stacked_bar}
\alias{cf_echarts_area}
\alias{cf_echarts_stacked_area}
\alias{cf_echarts_histogram}
\alias{cf_echarts_density}
\alias{cf_echarts_pie}
\alias{cf_echarts_donut}
\alias{cf_echarts_bubble}
\alias{cf_echarts_boxplot}
\alias{cf_echarts_heatmap}
\alias{cf_echarts_treemap}
\alias{cf_echarts_lollipop}
\alias{cf_apexcharts_scatter}
\alias{cf_apexcharts_line}
\alias{cf_apexcharts_spline}
\alias{cf_apexcharts_step}
\alias{cf_apexcharts_bar}
\alias{cf_apexcharts_stacked_bar}
\alias{cf_apexcharts_area}
\alias{cf_apexcharts_stacked_area}
\alias{cf_apexcharts_histogram}
\alias{cf_apexcharts_density}
\alias{cf_apexcharts_pie}
\alias{cf_apexcharts_donut}
\alias{cf_apexcharts_bubble}
\alias{cf_apexcharts_boxplot}
\alias{cf_apexcharts_heatmap}
\alias{cf_apexcharts_treemap}
\alias{cf_apexcharts_lollipop}
\alias{cf_googlecharts_scatter}
\alias{cf_googlecharts_line}
\alias{cf_googlecharts_spline}
\alias{cf_googlecharts_step}
\alias{cf_googlecharts_bar}
\alias{cf_googlecharts_stacked_bar}
\alias{cf_googlecharts_area}
\alias{cf_googlecharts_stacked_area}
\alias{cf_googlecharts_histogram}
\alias{cf_googlecharts_density}
\alias{cf_googlecharts_pie}
\alias{cf_googlecharts_donut}
\alias{cf_googlecharts_bubble}
\alias{cf_googlecharts_boxplot}
\alias{cf_googlecharts_heatmap}
\alias{cf_googlecharts_treemap}
\alias{cf_googlecharts_lollipop}
\alias{cf_amcharts_scatter}
\alias{cf_amcharts_line}
\alias{cf_amcharts_spline}
\alias{cf_amcharts_step}
\alias{cf_amcharts_bar}
\alias{cf_amcharts_stacked_bar}
\alias{cf_amcharts_area}
\alias{cf_amcharts_stacked_area}
\alias{cf_amcharts_histogram}
\alias{cf_amcharts_density}
\alias{cf_amcharts_pie}
\alias{cf_amcharts_donut}
\alias{cf_amcharts_bubble}
\alias{cf_amcharts_boxplot}
\alias{cf_amcharts_heatmap}
\alias{cf_amcharts_treemap}
\alias{cf_amcharts_lollipop}
\alias{cf_vchart_scatter}
\alias{cf_vchart_line}
\alias{cf_vchart_spline}
\alias{cf_vchart_step}
\alias{cf_vchart_bar}
\alias{cf_vchart_stacked_bar}
\alias{cf_vchart_area}
\alias{cf_vchart_stacked_area}
\alias{cf_vchart_histogram}
\alias{cf_vchart_density}
\alias{cf_vchart_pie}
\alias{cf_vchart_donut}
\alias{cf_vchart_bubble}
\alias{cf_vchart_boxplot}
\alias{cf_vchart_heatmap}
\alias{cf_vchart_treemap}
\alias{cf_vchart_lollipop}
\alias{cf_anychart_scatter}
\alias{cf_anychart_line}
\alias{cf_anychart_spline}
\alias{cf_anychart_step}
\alias{cf_anychart_bar}
\alias{cf_anychart_stacked_bar}
\alias{cf_anychart_area}
\alias{cf_anychart_stacked_area}
\alias{cf_anychart_histogram}
\alias{cf_anychart_density}
\alias{cf_anychart_pie}
\alias{cf_anychart_donut}
\alias{cf_anychart_bubble}
\alias{cf_anychart_boxplot}
\alias{cf_anychart_heatmap}
\alias{cf_anychart_treemap}
\alias{cf_anychart_lollipop}
\alias{cf_billboard_scatter}
\alias{cf_billboard_line}
\alias{cf_billboard_spline}
\alias{cf_billboard_step}
\alias{cf_billboard_bar}
\alias{cf_billboard_stacked_bar}
\alias{cf_billboard_area}
\alias{cf_billboard_stacked_area}
\alias{cf_billboard_histogram}
\alias{cf_billboard_density}
\alias{cf_billboard_pie}
\alias{cf_billboard_donut}
\alias{cf_billboard_bubble}
\alias{cf_billboard_boxplot}
\alias{cf_billboard_heatmap}
\alias{cf_billboard_treemap}
\alias{cf_billboard_lollipop}
\title{Renderer by chart-type wrapper functions}
\usage{
cf_plotly_scatter(data, x = NULL, y = NULL, group = NULL,
  label = NULL, sort_by = NULL, sort_desc = FALSE, ...,
  full_bundle = FALSE, embed = TRUE, minified = FALSE)
}
\arguments{
\item{data}{A data frame or matrix.}
\item{x, y, group}{Single column-name strings used for the selected chart mapping.}
\item{label}{A character vector of optional hover-detail columns. Mapped x, y, and group values are not repeated in the extra label text.}
\item{sort_by}{A character vector of columns used to sort the data before rendering.}
\item{sort_desc}{A logical value or vector controlling descending order for \code{sort_by}.}
\item{...}{Additional arguments passed to \code{cf_chart_html()}, including \code{renderer_style} and advanced native option lists.}
\item{full_bundle}{If \code{TRUE}, inline vendor JavaScript/CSS only where the renderer's terms permit it and retain the applicable notices. CDN-only renderers remain remotely linked and receive an explanatory legal annotation. If \code{FALSE}, emit CDN tags.}
\item{embed}{If \code{TRUE}, return an embeddable fragment. If \code{FALSE}, return a complete HTML document.}
\item{minified}{If \code{TRUE}, use compact markup and a minified renderer-by-chart-type runtime.}
}
\value{A character scalar containing an HTML fragment or document.}
\description{ChartForge creates one function for every supported renderer by chart-type combination. Function names follow renderer and chart-type identifiers, for example \code{cf_plotly_scatter()} and \code{cf_highcharts_heatmap()}.}
